% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2DNAStringSet.R
\name{bam2DNAStringSet}
\alias{bam2DNAStringSet}
\title{bam2DNAStringSet}
\usage{
bam2DNAStringSet(bamfile, refseq)
}
\arguments{
\item{bamfile}{bam file}

\item{refseq}{refseq, DNAStringSet object}
}
\value{
DNAStringSet object
}
\description{
convert bam file to aligned fasta file
}
\author{
Guangchuang Yu
}
