% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_align.R
\name{bs_align}
\alias{bs_align}
\title{bs_align}
\usage{
bs_align(x, method = "muscle", ...)
}
\arguments{
\item{x}{BStringSet or DNAStringSet}

\item{method}{alignment method}

\item{...}{additional parameters}
}
\value{
BStringSet
}
\description{
alignment
}
\author{
Guangchuang Yu
}

