% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln.R
\name{bs_aln}
\alias{bs_aln}
\title{bs_aln}
\usage{
bs_aln(x, method = "muscle", ...)
}
\arguments{
\item{x}{XStringSet object}

\item{method}{alignment method}

\item{...}{additional parameter}
}
\value{
aligned sequences, XStringSet object
}
\description{
sequence alignment
}
\author{
Guangchuang Yu
}
