% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{bs_filter}
\alias{bs_filter}
\title{bs_filter}
\usage{
bs_filter(x, pattern, by = "description", ignore.case = FALSE)
}
\arguments{
\item{x}{BStringSet object}

\item{pattern}{keyword for filter}

\item{by}{one of 'description' and 'sequence'}

\item{ignore.case}{logical}
}
\value{
BStringSet object
}
\description{
biological sequence filter by searching pattern
}
\author{
Guangchuang Yu
}

