% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{bs_rename}
\alias{bs_rename}
\title{bs_rename}
\usage{
bs_rename(x, mapping, position, sep, replace, prefix)
}
\arguments{
\item{x}{BStringSet object}

\item{mapping}{two column data.frame}

\item{position}{rename token at specific position}

\item{sep}{sepator to divide token}

\item{replace}{logical, whether replace token or not}

\item{prefix}{logical. If replace is FALSE, whether add to prefix or suffix}
}
\value{
BStringSet
}
\description{
rename sequence
}
\author{
Guangchuang Yu
}

