% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus}
\alias{consensus}
\title{consensus}
\usage{
consensus(x, type = "DNA")
}
\arguments{
\item{x}{XStringSet or XMultipleAlignment object}

\item{type}{currently, only DNA supported}
}
\value{
consensus sequence string
}
\description{
consensus of aligned sequences
}
\author{
Guangchuang Yu
}

