% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_combine.R
\name{fa_combine}
\alias{fa_combine}
\title{fa_combine}
\usage{
fa_combine(file1, file2, outfile = NULL, type = "interleaved")
}
\arguments{
\item{file1}{fasta file 1}

\item{file2}{fasta file 2}

\item{outfile}{output file}

\item{type}{one of interleaved and sequential}
}
\value{
BStringSet
}
\description{
combine 2 fasta files into 1
}
\author{
Guangchuang Yu
}
