% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_consensus.R
\name{fa_consensus}
\alias{fa_consensus}
\title{fa_consensus}
\usage{
fa_consensus(x, type = "DNA")
}
\arguments{
\item{x}{BStringSet object}

\item{type}{currently, only DNA supported}
}
\value{
consensus sequence string
}
\description{
consensus of aligned fasta file
}
\author{
Guangchuang Yu
}

