% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_conversion.R
\name{fa_to_interleaved}
\alias{fa_to_interleaved}
\alias{fa_to_sequential}
\title{fa_to_interleaved}
\usage{
fa_to_interleaved(file, outfile)

fa_to_sequential(file, outfile)
}
\arguments{
\item{file}{fasta file}

\item{outfile}{output file}
}
\value{
None

None
}
\description{
convert fasta file to interleaved format

convert fasta file to sequential format
}
\examples{
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
fa1 <- tempfile(fileext = '.fa')
fa2 <- tempfile(fileext = '.fa')
fa_to_interleaved(fa_file, fa1)
fa_to_sequential(fa_file, fa2)
}
\author{
Guangchuang Yu

Guangchuang Yu
}
