% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fa_read}
\alias{fa_read}
\title{fa_read}
\usage{
fa_read(file, type = "DNA")
}
\arguments{
\item{file}{fasta file}

\item{type}{one of 'DNA', 'RNA', 'AA' or 'UNKNOWN'}
}
\value{
BStringSet object
}
\description{
read fasta file
}
\author{
Guangchuang Yu
}
