% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{fa_rename}
\alias{fa_rename}
\title{fa_rename}
\usage{
fa_rename(fasfile, mapping_file, position, sep, mode, outfile)
}
\arguments{
\item{fasfile}{fasta file}

\item{mapping_file}{mapping file}

\item{position}{rename token at specific position}

\item{sep}{sepator to divide token}

\item{mode}{one of 'replace', 'prefix' or 'suffix'}

\item{outfile}{output file}
}
\value{
BStringSet object
}
\description{
rename fasta sequence name
}
\author{
Guangchuang Yu
}
