% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_conversion.R
\name{fa_to_sequential}
\alias{fa_to_sequential}
\title{fa_to_sequential}
\usage{
fa_to_sequential(file, outfile)
}
\arguments{
\item{file}{fasta file}

\item{outfile}{output file}
}
\description{
convert fasta file to sequential format
}
\author{
Guangchuang Yu
}

