% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{gb_read}
\alias{gb_read}
\title{gb_read}
\usage{
gb_read(file)
}
\arguments{
\item{file}{input genbank file}
}
\value{
sequence object
}
\description{
extract accession number and sequence from genbank file
}
\author{
Guangchuang Yu
}
