% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{get_id}
\usage{
get_id(x, sep = " ", position)
}
\arguments{
\item{x}{x}

\item{sep}{separator to split x}

\item{position}{id position}
}
\value{
id
}
\description{
get id at specific position
}
\author{
guangchuang yu
}

