% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega.r
\name{mega_read}
\alias{mega_read}
\title{mega_read}
\usage{
mega_read(file)
}
\arguments{
\item{file}{mega file}
}
\value{
BStringSet object
}
\description{
read sequence alignment file from MEGA
}
\examples{
mega_file <- system.file("extdata/mega/Crab_rRNA.meg", package="seqmagick")
mega_read(mega_file)
}
\author{
Guangchuang Yu
}
