% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega.r, R/read.R
\name{mega_read}
\alias{mega_read}
\alias{fa_read}
\alias{clw_read}
\alias{sth_read}
\title{read sequence alignment file}
\usage{
mega_read(file)

fa_read(file, type = "auto")

clw_read(file, type = "auto")

sth_read(file, type = "auto")
}
\arguments{
\item{file}{multiple sequence file}

\item{type}{one of 'DNA', 'RNA', 'AA', 'Protein', 'unknown' or 'auto'}
}
\value{
BStringSet object
}
\description{
read sequence alignment file
}
\examples{
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
fa_read(fa_file)

mega_file <- system.file("extdata/mega/Crab_rRNA.meg", package="seqmagick")
mega_read(mega_file)
}
\author{
Guangchuang Yu
}
