% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBI.R
\name{ncbi_fa_read}
\alias{ncbi_fa_read}
\title{ncbi_fa_read}
\usage{
ncbi_fa_read(acc, db = "nuccore", type = "auto", ...)
}
\arguments{
\item{acc}{accession number(s)}

\item{db}{supported db, currently 'nuccore'}

\item{type}{one of 'DNA', 'RNA', 'AA', 'unknown' or 'auto'}

\item{...}{additional parameters passed to downloader::download}
}
\value{
BStringSet object
}
\description{
download fasta sequence by accession number and read as sequence object
}
\examples{
\dontrun{
x <- ncbi_fa_read(c('AB115403', 'AB115403.1'))
x
}
}
\author{
Guangchuang Yu
}
