% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2fas.R
\name{phy2fas}
\alias{phy2fas}
\title{phy2fas}
\usage{
phy2fas(phyfile, outfile = "out.fas", type = "interleaved")
}
\arguments{
\item{phyfile}{phylip file}

\item{outfile}{output file}

\item{type}{one of interleaved and sequential}
}
\description{
convert phylip file to fasta file
}
\author{
Guangchuang Yu
}
