% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{phy_read}
\alias{phy_read}
\title{phy_read}
\usage{
phy_read(file)
}
\arguments{
\item{file}{phylip file}
}
\value{
BStringSet object
}
\description{
read aligned sequences in phylip format
}
\examples{
phy_file <- system.file("extdata/HA.phy", package="seqmagick")
phy_read(phy_file)
}
\author{
Guangchuang Yu
}
