% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameTXT.R
\name{renameTXT}
\alias{renameTXT}
\title{renameTXT}
\usage{
renameTXT(txt_file, name_file, sep = "_", split = TRUE)
}
\arguments{
\item{txt_file}{txt file}

\item{name_file}{name file}

\item{sep}{separator}

\item{split}{logical, split result or not}
}
\value{
None
}
\description{
rename txt file (eg Description line of fasta file)
according to first token (eg accession number)
}
\author{
Guangchuang Yu
}
