% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceInside.R
\name{replaceInside}
\alias{replaceInside}
\title{replaceInside}
\usage{
replaceInside(fasfile, from = "-", to = "N", outfile = NULL)
}
\arguments{
\item{fasfile}{fasta file}

\item{from}{character to be replaced, '-' by default}

\item{to}{replace 'from' to 'to', 'N' by default}

\item{outfile}{output file}
}
\value{
DNAStringSet
}
\description{
replace character for example from '-' to 'N' of fasta sequence that only applied inside sequence
any '-' character at start/end of the sequence (aligned seqs may contains '-' at prefix/suffix) will not be replaced
}
\author{
Guangchuang Yu
}
