% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmagick-package.R
\docType{package}
\name{seqmagick-package}
\alias{seqmagick}
\alias{seqmagick-package}
\title{seqmagick: Sequence Manipulation Utilities}
\description{
Tools for reading and writing biological sequences in multiple formats, including 'FASTA', 'PHYLIP', 'CLUSTAL', 'STOCKHOLM', 'MEGA' and 'GenBank'. Supports interleaved and sequential layouts where applicable, converts between formats, and manipulates sequence sets (e.g., filtering by patterns and computing consensus sequences from alignments). Also includes functions to download nucleotide records from NCBI by accession.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/seqmagick}
  \item Report bugs at \url{https://github.com/YuLab-SMU/seqmagick/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

}
\keyword{internal}
