% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{bs_consensus}
\alias{bs_consensus}
\title{bs_consensus}
\usage{
bs_consensus(x, type = "DNA", r = 1)
}
\arguments{
\item{x}{BStringSet object}

\item{type}{currently, only DNA supported}

\item{r}{if any NT > r, it will be selected as representative base}
}
\value{
consensus sequence string
}
\description{
consensus of aligned sequences
}
\author{
Guangchuang Yu
}
