% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming.R
\name{bs_hamming}
\alias{bs_hamming}
\title{bs_hamming}
\usage{
bs_hamming(x, count_indel = FALSE, ...)
}
\arguments{
\item{x}{BStringSet object}

\item{count_indel}{whether count indel or not}

\item{...}{additional parameters}
}
\value{
hamming distance
}
\description{
hamming distances of sequences
}
\examples{
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
x <- fa_read(fa_file)
## align first 5 sequences, use `bs_aln(x)` to align all sequences
aln <- bs_aln(x[1:5])
bs_hamming(aln)
}
\author{
Guangchuang Yu
}
