% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fa_read}
\alias{fa_read}
\title{fa_read}
\usage{
fa_read(file, type = "auto")
}
\arguments{
\item{file}{fasta file}

\item{type}{one of 'DNA', 'RNA', 'AA', 'unknown' or 'auto'}
}
\value{
BStringSet object
}
\description{
read fasta file
}
\examples{
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
fa_read(fa_file)
}
\author{
Guangchuang Yu
}
