% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_write.R
\name{fa_write}
\alias{fa_write}
\title{fa_write}
\usage{
fa_write(x, filepath, type = "interleaved")
}
\arguments{
\item{x}{XStringSet object}

\item{filepath}{output file}

\item{type}{one of interleaved and sequential}
}
\description{
write fasta file
}
\author{
Guangchuang Yu
}
\references{
\url{http://evolution.genetics.washington.edu/phylip/doc/sequence.html}
}

