% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{fa_write}
\alias{fa_write}
\title{fa_write}
\usage{
fa_write(x, outfile, type = "interleaved")
}
\arguments{
\item{x}{XStringSet object}

\item{outfile}{output file}

\item{type}{one of interleaved and sequential}
}
\value{
None
}
\description{
write fasta file
}
\examples{
phy_file <- system.file("extdata/HA.phy", package="seqmagick")
x <- phy_read(phy_file)
fa_file <- tempfile(fileext = '.fas')
fa_write(x, fa_file)
}
\references{
\url{http://evolution.genetics.washington.edu/phylip/doc/sequence.html}
}
\author{
Guangchuang Yu
}
