% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fas2phy.R
\name{fas2phy}
\alias{fas2phy}
\title{fas2phy}
\usage{
fas2phy(fasfile, outfile = "out.phy", type = "sequential")
}
\arguments{
\item{fasfile}{aligned sequences in fasta format}

\item{outfile}{output file}

\item{type}{one of interleaved and sequential}
}
\value{
None
}
\description{
convert fasta (aligned sequences) to phylip format
}
\author{
Guangchuang Yu
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
phy_file <- tempfile(fileext = ".phy")
fas2phy(fa_file, phy_file)
}
