% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{get_id}
\usage{
get_id(x, sep = " ", position)
}
\arguments{
\item{x}{sequence description line}

\item{sep}{separator to split x}

\item{position}{id position}
}
\value{
id
}
\description{
get id at specific position
}
\examples{
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
x <- fa_read(fa_file)
get_id(names(x)[1:5], sep = " ", position=1)
}
\author{
Guangchuang Yu
}
