% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{phy_write}
\alias{phy_write}
\title{phy_write}
\usage{
phy_write(x, outfile, type = "sequential")
}
\arguments{
\item{x}{XStringSet object}

\item{outfile}{output file}

\item{type}{one of interleaved and sequential}
}
\description{
write phylip file
}
\examples{
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
x <- phy_read(fa_file)
phy_file <- tempfile(fileext = '.phy')
phy_write(x, phy_file)
}
\author{
Guangchuang Yu
}
