% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{phy_write}
\alias{phy_write}
\title{phy_write}
\usage{
phy_write(x, outfile, type = "sequential")
}
\arguments{
\item{x}{XStringSet object}

\item{outfile}{output file}

\item{type}{one of interleaved and sequential}
}
\value{
None
}
\description{
write phylip file
}
\examples{
fa_file <- system.file("extdata/HA.fas", package="seqmagick")
x <- fa_read(fa_file)
aln <- bs_aln(x[1:5])
phy_file <- tempfile(fileext = '.phy')
phy_write(aln, phy_file)
}
\author{
Guangchuang Yu
}
