% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmagick-package.R
\docType{package}
\name{seqmagick-package}
\alias{seqmagick}
\alias{seqmagick-package}
\title{seqmagick: Sequence Manipulation Utilities}
\description{
Supports reading and writing sequences for different formats (currently interleaved and sequential formats for 'FASTA' and 'PHYLIP'), file conversion, and manipulation (e.g. filter sequences that contain specify pattern, export consensus sequence from an alignment).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YuLab-SMU/seqmagick}
  \item Report bugs at \url{https://github.com/YuLab-SMU/seqmagick/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

}
\keyword{internal}
